+++ crates/project-builder/src/project_builder.rs
@@ -38,7 +38,14 @@ pub struct ProjectBuilder<'app> {
     root: PathBuf,
 
     pub language: LanguageType,
-    pub toolchains: Vec<Id>,
+
+    // Toolchains that will be used as the fallback for tasks.
+    // These are filtered based on enabled.
+    pub toolchains_tasks: Vec<Id>,
+
+    // Toolchains that will be used for task/config inheritance.
+    // These are *not* filtered based on enabled.
+    pub toolchains_config: Vec<Id>,
 }
 
 impl<'app> ProjectBuilder<'app> {
@@ -63,7 +70,8 @@ impl<'app> ProjectBuilder<'app> {
             global_config: None,
             local_config: None,
             language: LanguageType::Unknown,
-            toolchains: vec![],
+            toolchains_tasks: vec![],
+            toolchains_config: vec![],
         })
     }
 
@@ -79,7 +87,7 @@ impl<'app> ProjectBuilder<'app> {
             .expect("Local config must be loaded before global config!");
 
         let global_config = tasks_manager.get_inherited_config(
-            &self.toolchains,
+            &self.toolchains_config,
             &local_config.stack,
             &local_config.type_of,
             &local_config.tags,
@@ -115,12 +123,14 @@ impl<'app> ProjectBuilder<'app> {
         };
 
         // Infer toolchains from language
-        if self.toolchains.is_empty() {
+        if self.toolchains_tasks.is_empty() {
             let mut added = false;
 
             if let Some(default_id) = &config.toolchain.default {
+                self.toolchains_config.push(default_id.to_owned());
+
                 if self.context.enabled_toolchains.contains(default_id) {
-                    self.toolchains.push(default_id.to_owned());
+                    self.toolchains_tasks.push(default_id.to_owned());
                     added = true;
                 }
             }
@@ -131,8 +141,10 @@ impl<'app> ProjectBuilder<'app> {
                 if let Some(platform) = &config.platform {
                     let default_id = platform.get_toolchain_id();
 
+                    self.toolchains_config.push(default_id.clone());
+
                     if self.context.enabled_toolchains.contains(&default_id) {
-                        self.toolchains.push(default_id);
+                        self.toolchains_tasks.push(default_id);
                         added = true;
                     }
                 }
@@ -146,18 +158,24 @@ impl<'app> ProjectBuilder<'app> {
             }
 
             if !added {
-                self.toolchains = detect_project_toolchains(
+                let toolchains = detect_project_toolchains(
                     self.context.workspace_root,
                     &self.root,
                     &self.language,
-                    self.context.enabled_toolchains,
                 );
+
+                self.toolchains_config.extend(toolchains.clone());
+
+                self.toolchains_tasks = toolchains
+                    .into_iter()
+                    .filter(|id| self.context.enabled_toolchains.contains(id))
+                    .collect();
             }
 
             trace!(
                 project_id = self.id.as_str(),
                 language = ?self.language,
-                toolchains = ?self.toolchains.iter().map(|tc| tc.as_str()).collect::<Vec<_>>(),
+                toolchains = ?self.toolchains_tasks.iter().map(|tc| tc.as_str()).collect::<Vec<_>>(),
                 "Unknown tasks toolchain, inferring from project language",
             );
         }
@@ -230,7 +248,12 @@ impl<'app> ProjectBuilder<'app> {
             language: self.language,
             root: self.root,
             source: self.source.to_owned(),
-            toolchains: self.toolchains,
+            // Should this be the config one?
+            toolchains: if self.toolchains_tasks.is_empty() {
+                vec![Id::raw("system")]
+            } else {
+                self.toolchains_tasks
+            },
             ..Project::default()
         };
 
@@ -372,7 +395,7 @@ impl<'app> ProjectBuilder<'app> {
         let mut tasks_builder = TasksBuilder::new(
             self.id,
             self.source,
-            &self.toolchains,
+            &self.toolchains_tasks,
             TasksBuilderContext {
                 enabled_toolchains: self.context.enabled_toolchains,
                 monorepo: self.context.monorepo,